gpwk/lib/python/requirements.txt
# Core dependencies
requests>=2.31.0
PyGithub>=2.1.1
python-dateutil>=2.8.2

# OpenTelemetry dependencies
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
opentelemetry-instrumentation-requests>=0.42b0

# Logging
structlog>=23.2.0

# CLI
click>=8.1.7

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Type checking
mypy>=1.7.0
types-requests>=2.31.0
